many: move helper to clientutil
