+++ overlord/hookstate/ctlcmd/set_test.go
@@ -509,49 +509,3 @@ func (s *registrySuite) TestRegistrySetExclamationMark(c *C) {
 `)
 	c.Check(stderr, IsNil)
 }
-
-func (s *registrySuite) TestParseConfigValues(c *C) {
-	// check basic setting and unsetting behaviour
-	confValues, err := ctlcmd.ParseConfigValues([]string{"foo=bar", "baz!"}, nil)
-	c.Assert(err, IsNil)
-	c.Assert(confValues, DeepEquals, map[string]interface{}{
-		"foo": "bar",
-		"baz": nil,
-	})
-
-	// parses JSON
-	opts := &ctlcmd.ParseConfigOptions{
-		Typed: true,
-	}
-	confValues, err = ctlcmd.ParseConfigValues([]string{`foo={"bar": 1}`}, opts)
-	c.Assert(err, IsNil)
-	c.Assert(confValues, DeepEquals, map[string]interface{}{
-		"foo": map[string]interface{}{
-			"bar": json.Number("1"),
-		},
-	})
-
-	// stores strings w/o parsing
-	opts.String = true
-	confValues, err = ctlcmd.ParseConfigValues([]string{`foo={"bar": 1}`}, opts)
-	c.Assert(err, IsNil)
-	c.Assert(confValues, DeepEquals, map[string]interface{}{
-		"foo": `{"bar": 1}`,
-	})
-
-	// default is to parse
-	confValues, err = ctlcmd.ParseConfigValues([]string{`foo={"bar": 1}`}, nil)
-	c.Assert(err, IsNil)
-	c.Assert(confValues, DeepEquals, map[string]interface{}{
-		"foo": map[string]interface{}{
-			"bar": json.Number("1"),
-		},
-	})
-
-	// unless it's not valid JSON
-	confValues, err = ctlcmd.ParseConfigValues([]string{`foo={"bar": 1`}, nil)
-	c.Assert(err, IsNil)
-	c.Assert(confValues, DeepEquals, map[string]interface{}{
-		"foo": `{"bar": 1`,
-	})
-}
